Revert "Bugfix FXIOS-5660 [v110] Continuation crash (#13092)"
This reverts commit bb51425c317008ce901f22738b4e9c818de78602.
